+++ experimental/puppeteer-firefox/misc/puppeteer.cfg
@@ -137,6 +137,9 @@ pref(
   "http://%(server)s/dummy/discoveryURL",
 );
 
+pref("extensions.screenshots.disabled", true);
+pref("extensions.screenshots.upload-disabled", true);
+
 // Allow the application to have focus even it runs in the background
 pref("focusmanager.testmode", true);
